refactor(app): 收口应用层共享类型边界
Why:
- 让 `app/` 继续从“新入口”收口成真实分层，而不只是把旧文件原样搬家
- 提前把共享类型与 remote dataclass 独立出来，降低后续清理兼容层时的耦合度

What:
- 将 `AppContext` 相关共享类型下沉到 `app/types.py`，让 `app/context.py` 聚焦构建逻辑
- 将 remote dataclass 下沉到 `app/remote.py`，并让 `app/facade.py` 只保留编排逻辑
- 将仓库内部仅依赖类型的调用点切到 `app.types`，保留 `create_app_context` 在 `app.context`

Risk:
- 兼容性影响：无对外新增破坏，旧入口仍保留兼容层，内部导入路径进一步向 `app.*` 收敛
- 潜在风险：若后续还有未覆盖代码依赖旧模块里的实现位置，最终删兼容层前仍需继续扫尾

Test:
- `uv run ruff check src/lol_audio_unpack/app src/lol_audio_unpack/app_context.py src/lol_audio_unpack/facade.py src/lol_audio_unpack/__init__.py src/lol_audio_unpack/cli src/lol_audio_unpack/gui src/lol_audio_unpack/manager src/lol_audio_unpack/mapping src/lol_audio_unpack/unpack src/lol_audio_unpack/runtime/remote/preparer.py src/lol_audio_unpack/model.py scripts/benchmark_cli.py scripts/benchmark_remote_live.py tests/conftest.py tests/test_app_package_api.py tests/test_app_context.py tests/test_champion_visibility.py tests/test_remote_preparer.py tests/test_cli.py tests/test_cli_invocation.py tests/test_gui_config.py tests/test_shared_data_controller.py tests/test_task_runner.py tests/test_task_creation_card.py tests/test_context_injection.py`
- `uv run pytest -q tests/test_app_package_api.py tests/test_app_context.py tests/test_context_injection.py tests/test_champion_visibility.py tests/test_remote_preparer.py tests/test_cli.py tests/test_cli_invocation.py tests/test_gui_config.py tests/test_shared_data_controller.py tests/test_task_runner.py tests/test_task_creation_card.py`
